feat: withUrl 딜레이 추가 및 run 메서드 상위 클래스로 이동
